refactor(entities): extract EntityForgetService — split read vs forget (max-params 2/3)
EntitiesService injected 4 deps (surreal, config, metrics, embedder). The
read path (getProfile/timeline/connections/freshnessWatermark) uses only
surreal; every other dep belonged to the GDPR erasure path. Move forget()
— cascade delete + PII-mirror purge + HMAC tombstone + embedder-cache
eviction — into EntityForgetService (surreal, config, embedder). The
public forget() on EntitiesService now delegates and counts the metric,
leaving it at 3 deps (surreal, forgetService, metrics).

Part of the max-params=3 god-class split program.
